requirements.txt
blinker==1.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
filelock==3.9.0
Flask==3.0.0
fsspec==2023.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
numpy==1.26.1
Pillow==9.3.0
requests==2.28.1
sympy==1.12
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
triton==2.1.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
